Conflicts:
	src/main/java/com/divudi/bean/pharmacy/PharmacyBillSearch.java
	src/main/java/com/divudi/bean/pharmacy/PharmacyItemExcelManager.java
	src/main/java/com/divudi/entity/BillFee.java
	src/main/resources/META-INF/persistence.xml
	src/main/setup/glassfish-resources.xml
	src/main/webapp/inward_admission.xhtml
	src/main/webapp/pharmacy_bill_dealor.xhtml
	src/main/webapp/resources/template/template.xhtml

+++ src/main/java/com/divudi/bean/pharmacy/PharmacyIssueController.java
@@ -458,7 +458,7 @@ public class PharmacyIssueController implements Serializable {
         tbi.setCreatedAt(Calendar.getInstance().getTime());
         tbi.setCreater(getSessionController().getLoggedUser());
         PharmaceuticalBillItem tmpPharmacyItem = tbi.getPharmaceuticalBillItem();
-        tmpPharmacyItem.setQty((float) (0 - tbi.getQty()));
+        tmpPharmacyItem.setQty((double) (0 - tbi.getQty()));
         tbi.setPharmaceuticalBillItem(null);
         getBillItemFacade().create(tbi);
 
@@ -642,7 +642,7 @@ public class PharmacyIssueController implements Serializable {
         billItem.getPharmaceuticalBillItem().setDoe(getStock().getItemBatch().getDateOfExpire());
         billItem.getPharmaceuticalBillItem().setFreeQty(0.0f);
         billItem.getPharmaceuticalBillItem().setItemBatch(getStock().getItemBatch());
-        billItem.getPharmaceuticalBillItem().setQty((float) (double) qty);
+        billItem.getPharmaceuticalBillItem().setQty((double) (double) qty);
 
         //Rates
         //Values
